Move `split_subtree` into Dom
